moved modus package and ModusBuffer class
* moved the modus package to be within the app package. reasoning: the modus classes are tied close enough to the app package and won't exist elsewhere.
* moved the ModusBuffer class to the model package. reasoning: the class acts as a buffer of data and is responsable for holding application state, so being a model for the modus classes makes sense.
