+++ IS_Java_Version/IS_class_diagram.class.violet.html
@@ -422,7 +422,7 @@ Modus</text>
       <borderColor reference="12"/>
       <textColor reference="12"/>
       <name id="123" justification="1" size="3" underlined="false">
-        <text>modus</text>
+        <text>app.modus</text>
       </name>
       <content id="124" justification="1" size="4" underlined="false">
         <text></text>
