+++ IS_Java_Version/src/app/controller/ModusManager.java
@@ -14,28 +14,29 @@ import java.util.stream.Collectors;
 
 import app.model.Card;
 import app.model.Metadata;
+import app.model.ModusBuffer;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.StackPane;
-import modus.*;
+import app.modus.*;
 
 /**
- * The ModusManager class provides support to the sylladex in managing the modi in the modus package. This enables users
- * to take a valid and {@link modus.Modus Modus} conforming java class and "plug" it into the sylladex for use. <br><br>
- * This class also tracks what modus is currently in use, and what modi are currently available, as well as metadata
- * about the particular modus, to help let the sylladex "know" how to handle certain events and actions.
+ * The ModusManager class provides support to the sylladex in managing the modi in the app.modus package. This enables users
+ * to take a valid and {@link app.modus.Modus Modus} conforming java class and "plug" it into the sylladex for use. <br><br>
+ * This class also tracks what app.modus is currently in use, and what modi are currently available, as well as metadata
+ * about the particular app.modus, to help let the sylladex "know" how to handle certain events and actions.
  * <p> Author's note: originally the direction of the Modus Manager was to allow users to add
  * a java file that complied with Modus interface to a file directory and then the sylladex could be refreshed and load
- * the code given, but this form of plug-and-play of adding "valid" modus java files to to a given directory for the
+ * the code given, but this form of plug-and-play of adding "valid" app.modus java files to to a given directory for the
  * manager to pick up for the sylladex to use is a bit too insecure for my preference. it would be better to simply
- * allow for a person to mod the open-source version of this by adding those java files mentioned above to the modus
+ * allow for a person to mod the open-source version of this by adding those java files mentioned above to the app.modus
  * package and then compile it into a jar. Later, once the modusValidation function can be more robust and check for
  * malicious behavior, then this direction can be resumed and utilize the files through a URL filestream from
  * getResourcesAsStream to temp files for the jar to use until the sylladex is closed, or use file chooser which would
- * allow the jar to keep track of the various modus files wherever they happen to reside. would require re- enabling the
+ * allow the jar to keep track of the various app.modus files wherever they happen to reside. would require re- enabling the
  * tracking of File paths through Metadata class.
  *
  * @author Triston Scallan
- * @see modus.Modus
+ * @see app.modus.Modus
  */
 class ModusManager {
     private ModusBuffer                  modusBuffer;
@@ -58,10 +59,10 @@ class ModusManager {
                                            wrappedDeck,
                                            wrappedOpenHand);
 
-        //populate a list with the class names in the modus package
+        //populate a list with the class names in the app.modus package
         List<String> modusNameList = createClassNameList();
 
-        //convert class names to class objects, filter out invalid modus classes, and collect
+        //convert class names to class objects, filter out invalid app.modus classes, and collect
         modusClassList = modusNameList.stream()
                                       //convert the names into class objects
                                       .map(className -> {
@@ -99,7 +100,7 @@ class ModusManager {
 
     /**
      * @param modusClass
-     *         the modus to update the selection to
+     *         the app.modus to update the selection to
      */
     <T extends Modus> void updateCurrentModus(Class<T> modusClass) throws IllegalAccessException, InstantiationException, IllegalArgumentException {
         //instantiate the desired class and set it to this#currentModusMetadata
@@ -110,7 +111,7 @@ class ModusManager {
         //replace old instance so it may be GC'd
         currentModusMetadata = clazzInstance.getMETADATA();
 
-        //reset previous modus specific data in modusBuffer
+        //reset previous app.modus specific data in modusBuffer
         modusBuffer.clearModusInput();
         modusBuffer.clearModusInputRedirector();
     }
@@ -152,10 +153,10 @@ class ModusManager {
     }
 
     /**
-     * Messages a modus command statement to the currently selected modus for execution
+     * Messages a app.modus command statement to the currently selected app.modus for execution
      *
      * @param command
-     *         a modus command associated with a ModusCommandMap entry
+     *         a app.modus command associated with a ModusCommandMap entry
      * @param args
      *         the arguments to supply the command with
      */
@@ -184,9 +185,9 @@ class ModusManager {
     }
 
     /**
-     * Initializes a new object of the current modus class, then effectively replaces the old object's reference.
+     * Initializes a new object of the current app.modus class, then effectively replaces the old object's reference.
      *
-     * @throws RuntimeException if the modus' constructor cannot be accessed or fails
+     * @throws RuntimeException if the app.modus' constructor cannot be accessed or fails
      */
     void resetModus() throws RuntimeException {
         try {
@@ -195,14 +196,14 @@ class ModusManager {
             throw new RuntimeException("Access to " + getCurrentModus().REFERENCE.getClass().getSimpleName() + " constructor was prevented.",
                                        e);
         } catch (IllegalArgumentException | InstantiationException e) {
-            throw new RuntimeException(getCurrentModus().REFERENCE.getClass().getSimpleName() + " modus constructor failed.", e);
+            throw new RuntimeException(getCurrentModus().REFERENCE.getClass().getSimpleName() + " app.modus constructor failed.", e);
         }
     }
 
     /**
-     * gets a list of all the modus class names. This is done in a modular and generalized way utilizing reflection.
-     * First it will attempt to get the name of the modus package. after acquiring the name it will then convert it into
-     * a resource to find the URL and URI path of the modus package so that it can convert it to a directory. it will
+     * gets a list of all the app.modus class names. This is done in a modular and generalized way utilizing reflection.
+     * First it will attempt to get the name of the app.modus package. after acquiring the name it will then convert it into
+     * a resource to find the URL and URI path of the app.modus package so that it can convert it to a directory. it will
      * then iterate through the directory and create a list of the class files contained. this will first be performed
      * for a file system and then for a jarfile system.
      */
@@ -275,10 +276,10 @@ class ModusManager {
      * Validates a class as an instantiable subclass of Modus
      *
      * @param modusClass
-     *         the modus class literal to be drawn from
+     *         the app.modus class literal to be drawn from
      * @param <M>
      *         a subclass of Modus
-     * @return True if modus can have a running instance created via reflection.
+     * @return True if app.modus can have a running instance created via reflection.
      */
     private <M extends Modus> boolean validateModusFile(Class<M> modusClass) {
         if (modusClass == null) return false;
