+++ IS_Java_Version/src/app/model/Metadata.java
@@ -1,13 +1,13 @@
 package app.model;
 
 import app.util.ModusCommandMap;
-import modus.Modus;
+import app.modus.Modus;
 
 /**
  * The Metadata class is a container for information about a Modus' functionality
  * 	and abilities.
  * <p> Contains the NAME (String), REFERENCE (Modus), COMMAND_MAP (String, Integer),
- * 	and modusFile (File) of a given modus java file.
+ * 	and modusFile (File) of a given app.modus java file.
  * @author Triston Scallan
  *
  */
@@ -21,7 +21,7 @@ public class Metadata {
 	 */
 	public final Modus           REFERENCE;
 	/**
-	 * a mapping between a modus' command name and the associated command lambda.
+	 * a mapping between a app.modus' command name and the associated command lambda.
 	 */
 	public final ModusCommandMap COMMAND_MAP;
 	/**
