+++ IS_Java_Version/src/app/modus/Heap.java
@@ -1,4 +1,4 @@
-package modus;
+package app.modus;
 
 import java.util.LinkedHashMap;
 
@@ -12,7 +12,7 @@ import app.model.Metadata;
 public class Heap implements Modus {
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#entry(int, java.lang.Object[])
+	 * @see app.modus.Modus#entry(int, java.lang.Object[])
 	 */
 	@Override
 	public String entry(int functionCode, String... objects) {
@@ -20,7 +20,7 @@ public class Heap implements Modus {
 	}
 	
 	/* (non-Javadoc)
-	 * @see modus.Modus#createFunctionMap()
+	 * @see app.modus.Modus#createFunctionMap()
 	 */
 	@Override
 	public LinkedHashMap<String, Integer> createFunctionMap() {
@@ -28,21 +28,21 @@ public class Heap implements Modus {
 	}
 	
 	/* (non-Javadoc)
-	 * @see modus.Modus#save()
+	 * @see app.modus.Modus#save()
 	 */
 	@Override
 	public void save() {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#load(int)
+	 * @see app.modus.Modus#load(int)
 	 */
 	@Override
 	public void load(int mode) throws Exception {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#capture(java.lang.String)
+	 * @see app.modus.Modus#capture(java.lang.String)
 	 */
 	@Override
 	public Boolean capture(String item) {
@@ -50,7 +50,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#addCard(app.model.Card)
+	 * @see app.modus.Modus#addCard(app.model.Card)
 	 */
 	@Override
 	public Boolean addCard(Card card) {
@@ -58,7 +58,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#takeOutCard(java.lang.Object[])
+	 * @see app.modus.Modus#takeOutCard(java.lang.Object[])
 	 */
 	@Override
 	public Card takeOutCard(Object... objects) {
@@ -66,7 +66,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#isFull()
+	 * @see app.modus.Modus#isFull()
 	 */
 	@Override
 	public Boolean isFull() {
@@ -74,7 +74,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#isEmpty()
+	 * @see app.modus.Modus#isEmpty()
 	 */
 	@Override
 	public Boolean isEmpty() {
@@ -82,7 +82,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#findItemName(java.lang.String)
+	 * @see app.modus.Modus#findItemName(java.lang.String)
 	 */
 	@Override
 	public String findItemName(String givenName) {
@@ -90,14 +90,14 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#drawToDisplay()
+	 * @see app.modus.Modus#drawToDisplay()
 	 */
 	@Override
 	public void drawToDisplay() {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#getMETADATA()
+	 * @see app.modus.Modus#getMETADATA()
 	 */
 	@Override
 	public Metadata getMETADATA() {
@@ -105,7 +105,7 @@ public class Heap implements Modus {
 	}
 
 	/* (non-Javadoc)
-	 * @see modus.Modus#description()
+	 * @see app.modus.Modus#description()
 	 */
 	@Override
 	public String description() {
